kafka-consumer-worker/requirements.txt
kafka-python==2.0.2
aiokafka==v0.10.0
aiomysql==0.2.0